Spring cleaning following full tabs migration

+++ src/main/java/com/cellterminal/integration/AE2WUTIntegration.java
@@ -127,7 +127,7 @@ public class AE2WUTIntegration {
     private static void registerRecipeIngredientInternal() {
         // Add to the ingredient list so WUT recognizes our terminal in crafting
         ItemStack ingredient = new ItemStack(ItemRegistry.WIRELESS_CELL_TERMINAL);
-        AllWUTRecipe.itemList.put((int) getCellTerminalMode(), ingredient);
+        AllWUTRecipe.itemList.put(getCellTerminalMode(), ingredient);
     }
 
     /**
@@ -233,7 +233,7 @@ public class AE2WUTIntegration {
     @Optional.Method(modid = MODID)
     private static ItemStack getWUTModeIconInternal(byte mode) {
         // Use AllWUTRecipe.itemList which has the actual terminal ItemStacks for all modes
-        ItemStack icon = AllWUTRecipe.itemList.get((int) mode);
+        ItemStack icon = AllWUTRecipe.itemList.get(mode);
 
         return icon != null ? icon : ItemStack.EMPTY;
     }
